fix(search): stop indexing canvas content
