+++ services/search_processing_service/src/parsers/mod.rs
@@ -1,4 +1,3 @@
-pub mod canvas;
 pub mod docx;
 pub mod markdown;
 #[cfg(feature = "pdf")]
